requirements.txt
anthropic>=0.40.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.13.0
python-dotenv>=1.0.0
streamlit>=1.35.0
